% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{.validate_square_matrix}
\alias{.validate_square_matrix}
\title{Validate a square matrix}
\usage{
.validate_square_matrix(M, p, nm = "matrix")
}
\arguments{
\item{M}{Matrix.}

\item{p}{Expected dimension.}

\item{nm}{Object name for messages.}
}
\description{
Validate a square matrix
}
\keyword{internal}
